# Compare path component by component in match/route.go