$ git switch master
Switched to branch 'master'
Your branch is up to date with 'origin/master'.
$ go test -v -run=^$ -bench BenchmarkRouters/match -benchmem | tee master.txt
goos: linux
goarch: amd64
pkg: github.com/benhoyt/go-routing
cpu: AMD Ryzen 7 PRO 4750GE with Radeon Graphics
BenchmarkRouters
BenchmarkRouters/match
BenchmarkRouters/match-16                 727123              1786 ns/op             624 B/op          9 allocs/op
PASS
ok      github.com/benhoyt/go-routing   1.326s
$ git switch match_by_path_component
Switched to branch 'match_by_path_component'
$ go test -v -run=^$ -bench BenchmarkRouters/match -benchmem | tee
match_by_component.txt
goos: linux
goarch: amd64
pkg: github.com/benhoyt/go-routing
cpu: AMD Ryzen 7 PRO 4750GE with Radeon Graphics
BenchmarkRouters
BenchmarkRouters/match
BenchmarkRouters/match-16                 740457              1916 ns/op             624 B/op          9 allocs/op
PASS
ok      github.com/benhoyt/go-routing   1.445s
$ benchstat master.txt match_by_component.txt
name              old time/op    new time/op    delta
Routers/match-16    1.79µs ± 0%    1.92µs ± 0%   ~     (p=1.000 n=1+1)

name              old alloc/op   new alloc/op   delta
Routers/match-16      624B ± 0%      624B ± 0%   ~     (all equal)

name              old allocs/op  new allocs/op  delta
Routers/match-16      9.00 ± 0%      9.00 ± 0%   ~     (all equal)